Fix sizing bug on restart
When there are writes dispatched in quick succession and a later one
returns the state to an earlier state, it can look like no write is
necessary if we are looking at the stale earlier state.

This was a problem when restarting the window manager. With some of the
windows already in the correct place, during app discovery a couple
writes would be issued in quick succession, the last returning some
windows to their original place.

To solve this we compare to the frame we last wrote, rather than the
last read frame.