Fix issue where raising an app raises all windows
Unfortunately this commit removes our "eventual consistency" guarantee
(which was only guaranteed if nothing timed out). Now we can have raise
requests for different apps race with one another once they are issued
by our wm.

But fear not, it should be possible to gain it back in the future!