Fix a race in run_loop test
The test it_can_wake_from_multiple_threads checked the number of wakeups
after reading from the handler channel, but the callback in
spawn_run_loop_thread only incremented the number of wakeups after
sending on the channel. This likely only worked because the other thread
was usually too slow to wake up to see the old number of wakeups.