Fix window not being selected on startup
Don't return a value from Reactor::main_window() if the window hasn't
been discovered yet. This causes us to delay the
LayoutEvent::WindowRaised until the window has actually been added to
the layout.

Also simplify how we keep track of the main window and recompute it
instead of manually managing a state machine. If we can afford to
recompute the window layout each step, we can definitely afford to
compute this.

Fixes #2.