Only save layouts when the user modifies them
This ensures that we don't have "junk" layouts saved for transient
screen configurations.

The login screen is a cache for these since it can display a different
resolution. After logging in on a new screen configuration we would
sometimes clone the layout for the login screen size rather than the one
the user was actually using before.