Implement focus_follows_mouse
Start using make_key_window every time a window is raised to ensure it
has focus. This is a different property from is_frontmost and we don't
cleanly separate the concepts today. macOS seems to make it very hard to
even know what window has focus.

This involved some trickery to avoid deadlocking on reactor's hacky
blocking_recv() when raising a window. If we raised a system dialog
window (iTerm hotkey window in my testing) we would never get an
ApplicationActivated event. In fact, no events seem to be firing in the
system at all when switching between a dialog like this and another
window.

For now we mark the raise as complete without waiting for an event; for
nonstandard windows that are not in the layout, it shouldn't matter.