Combine focus_window with other windows in the same app/screen
This can be a significant optimization that comes up in everyday use
because it saves us from potentially having to raise the app a second
time, and always saves some round trips to the window server and app.

In my testing raising two Zed windows, it saves 20ms on average.